requirements.txt
Flask
requests
cassandra-driver
flask_sqlalchemy